fixup "test/cache: refine test pattern" (x86_32)
genodelabs/genode#4454

+++ repos/os/src/test/cache/genode_time.h
@@ -31,7 +31,7 @@ struct Time
 		Timestamp diff = t2._ts - t1._ts;
 		if (t2._ts < t1._ts) diff--;
 
-		return Duration { diff };
+		return Duration { (unsigned long)diff };
 	}
 };
